Make `datef` POSIX compatible
`datef` had a huge caveat of not being POSIX compliant, which prevented
it's usage for example in container environments like `alpine`.

Refactor and improve the date input parsing logic. Use default
assignments to variables, which decreased the complexity of the logic.
Remove usage of arrays to store user given input dates and instead just
use two variables `date_a` and `date_b`. The maximum input dates is 2.

The first given input date is always the one compared to.

For example
Given: `2023-02-12`, is compared to now.
Given: `2023-02-12` and `2023-02-13`. The first one is compared to the
second one and outputs `1 day ago`.

Run shellcheck and shfmt, fix accordingly.