Add support for GNU date util
In Darwin and Linux the date utilities are different and the commands
they take vary quite a bit. Add support to set correct flags accordingly
if running in Darwin or Linux system.

Previously using `datef` in _any_ other system than Darwin would result
in an error due to incorrect flags.

`datef` can now be used for example in alpine container with dash.